# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Treinamento API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/hadassa/categories": {
            "get": {
                "tags": [
                    "Hadassa Categories"
                ],
                "summary": "List all HadassaCategorys",
                "operationId": "15d96450af2d4c0c7e04a2e24e2e103e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hadassa Categories"
                ],
                "summary": "Create a new HadassaCategory",
                "operationId": "c55589ffb8e1a692065fe72aedfc4b78",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categoria": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/hadassa/categories/{id}": {
            "get": {
                "tags": [
                    "Hadassa Categories"
                ],
                "summary": "Get HadassaCategory by ID",
                "operationId": "26b5709d2a069244d1ff6f8b7760bbcd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hadassa Categories"
                ],
                "summary": "Update a HadassaCategory",
                "operationId": "7c576aa2bf7f7ccb9c0edb0aef0063d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categoria": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hadassa Categories"
                ],
                "summary": "Delete a HadassaCategory",
                "operationId": "4cf7c56e67c601454f9d0cee8b804673",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/hadassa/posts": {
            "get": {
                "tags": [
                    "Hadassa Posts"
                ],
                "summary": "List all HadassaPosts",
                "operationId": "47326627ad93f923fa5f8c2cc3ff7a9c",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hadassa Posts"
                ],
                "summary": "Create a new HadassaPost",
                "operationId": "fab31f6a9f4fdad77f7d93c8c496253c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "id_categoria": {
                                        "type": "string"
                                    },
                                    "imagem": {
                                        "type": "string"
                                    },
                                    "author_id": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/hadassa/posts/{id}": {
            "get": {
                "tags": [
                    "Hadassa Posts"
                ],
                "summary": "Get HadassaPost by ID",
                "operationId": "34530a559e2cad5ef40c107ef8bdc7a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hadassa Posts"
                ],
                "summary": "Update a HadassaPost",
                "operationId": "c02a548b7ee453eaac6702a4b0c8d70e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    },
                                    "id_categoria": {
                                        "type": "string"
                                    },
                                    "imagem": {
                                        "type": "string"
                                    },
                                    "author_id": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hadassa Posts"
                ],
                "summary": "Delete a HadassaPost",
                "operationId": "2f0e65602ce80fe3e53d5f8018b88f88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
}